Add files via upload
This project builds a machine learning model to predict multiple water pollutant levels based on the station ID and year. The data was cleaned by removing missing values, and station IDs were one-hot encoded. A Random Forest regressor wrapped in a MultiOutputRegressor was trained to predict six pollutants simultaneously. The model was evaluated using MSE and R² scores and saved for future predictions.